# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Severinno Marketplace API",
    "version": "0.2.0",
    "description": "API do Marketplace de serviços com geolocalização"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/api/bookings/contract": {
      "get": {
        "summary": "Retrieve",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Bookings"
        ]
      },
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/emergency-dispatch": {
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/provider/routes/optimize-daily": {
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Provider"
        ]
      }
    },
...
    "/api/provider/wallet/withdraw": {
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Provider"
        ]
      }
    },
...
    "/api/webhooks/sentry-alert": {
      "post": {
        "summary": "POST /api/webhooks/sentry-alert — Recebe alertas do GlitchTip/Sentry e encaminha",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    }
  },
...
}